requirements.txt
google-generativeai
nltk
python-dotenv
spacy
